Fix compilation warnings in JS and Windows builds
Warnings raised by Emscripten 1.38.0 and MinGW64 5.0.4 / GCC 8.3.0.

JS can now build with `werror=yes warnings=extra`.
MinGW64 still has a few warnings to resolve with `warnings=extra`,
and only one with `warnings=all`.

Part of #29033 and #29801.

+++ platform/windows/power_windows.cpp
@@ -89,7 +89,7 @@ bool PowerWindows::GetPowerInfo_Windows() {
 		if (pct != 255) { /* 255 == unknown */
 			percent_left = (pct > 100) ? 100 : pct; /* clamp between 0%, 100% */
 		}
-		if (secs != 0xFFFFFFFF) { /* ((DWORD)-1) == unknown */
+		if (secs != (int)0xFFFFFFFF) { /* ((DWORD)-1) == unknown */
 			nsecs_left = secs;
 		}
 	}
